Modify file(class) Employee
